+++ lib/bluez.dart
@@ -1,2 +1,14 @@
+export 'src/bluez_adapter.dart';
+export 'src/bluez_battery.dart';
+export 'src/bluez_manufacturer_id.dart';
+export 'src/bluez_agent.dart';
+export 'src/bluez_characteristic.dart';
 export 'src/bluez_client.dart';
+export 'src/bluez_device.dart';
+export 'src/bluez_enums.dart';
+export 'src/bluez_exception.dart';
+export 'src/bluez_gatt_descriptor.dart';
+export 'src/bluez_gatt_service.dart';
+export 'src/bluez_object.dart';
 export 'src/bluez_uuid.dart';
+export 'src/bluez_advertisement.dart';
